requirements.txt
# gee_service/requirements.txt
Flask
flask-cors
earthengine-api
python-dotenv
gunicorn  # Añade esto, ya que Render lo necesita para la producción.